modify: separate dirextory && make Malefile
