+++ .gitignore
@@ -1,1 +1,3 @@
-a.out
+*.out
+*.o
+*.a
